Allows remote jobs that could be done from anywhere
Someone brought up that a literal reading could mean a job that said literally "work from anywhere" was contrary to the former version of the rules. This version shifts a "is" to "can" allowing the category of jobs that allow work anywhere. 

I do not believe we intended to exclude them, and reception to this possible change seems positive.