Adds who #jobs and #gigs are for
There are some positions which are in the fuzzy line between temporary jobs and permanent jobs. They often do have endgames though, where the worker will be forced to take W2 employment or leave. 

This PR attempts to clarify that. 

It also clarifies a long standing complaint about remote roles.